add more dialects, move base table_data to separate module
