+++ tests/dialects/test_hql.py
@@ -138,7 +138,7 @@ def test_partitioned_by_hql_output_mode_hql():
             "schema": "database",
             "table_name": "table_name",
             "tablespace": None,
-            "location": None,
+            "temp": False,
             "row_format": None,
             "fields_terminated_by": None,
             "collection_items_terminated_by": None,
@@ -249,7 +249,6 @@ def test_partitioned_by_hql_output_mode_hql():
             "stored_as": None,
             "table_name": "table_name2",
             "tablespace": None,
-            "location": None,
             "row_format": None,
             "fields_terminated_by": None,
             "collection_items_terminated_by": None,
@@ -384,7 +383,6 @@ def test_stored_as_hql_showed():
             "table_name": "table_name",
             "tablespace": None,
             "stored_as": "PARQUET",
-            "location": None,
             "row_format": None,
             "fields_terminated_by": None,
             "collection_items_terminated_by": None,
@@ -730,7 +728,6 @@ def test_hql_row_format():
             "index": [],
             "partitioned_by": [],
             "stored_as": "TEXTFILE",
-            "location": None,
             "external": False,
             "schema": "default",
             "table_name": "salesorderdetail",
@@ -810,7 +807,6 @@ def test_fields_terminated_by_hql():
             "index": [],
             "partitioned_by": [],
             "stored_as": "TEXTFILE",
-            "location": None,
             "row_format": "DELIMITED",
             "fields_terminated_by": "','",
             "external": False,
@@ -891,7 +887,6 @@ def test_collection_items_terminated_by_hql():
             "index": [],
             "partitioned_by": [],
             "stored_as": "TEXTFILE",
-            "location": None,
             "row_format": "DELIMITED",
             "fields_terminated_by": "'\\002'",
             "external": False,
@@ -974,7 +969,6 @@ def test_map_keys_terminated_by_hql():
             "index": [],
             "partitioned_by": [],
             "stored_as": "TEXTFILE",
-            "location": None,
             "row_format": "DELIMITED",
             "fields_terminated_by": "','",
             "external": False,
@@ -1019,7 +1013,6 @@ def simple_structure_type_support():
             "external": False,
             "fields_terminated_by": None,
             "index": [],
-            "location": None,
             "map_keys_terminated_by": None,
             "partitioned_by": [],
             "primary_key": [],
@@ -1854,7 +1847,6 @@ def test_skewed_by():
             "partitioned_by": [],
             "tablespace": None,
             "stored_as": "DIRECTORIES",
-            "location": None,
             "row_format": None,
             "fields_terminated_by": None,
             "lines_terminated_by": None,
@@ -2193,7 +2185,6 @@ WITH SERDEPROPERTIES ( 'key1'='value1', 'key2'='value2' , 'key3'='value3' , 'key
             "fields_terminated_by": None,
             "index": [],
             "lines_terminated_by": None,
-            "location": None,
             "map_keys_terminated_by": None,
             "partitioned_by": [],
             "primary_key": [],
