+++ tests/non_statement_tests/test_args.py
@@ -14,13 +14,11 @@ def test_json_dump_arg():
         '"unique": false, "nullable": true, "default": null, "check": null}, {"name": "value", '
         '"type": "STRING", "size": null, "references": null, "unique": false, "nullable": true, '
         '"default": null, "check": null}], "primary_key": [], "alter": {}, "checks": [], "index": [], '
-        '"partitioned_by": [], "tablespace": null, "stored_as": "DIRECTORIES", "location": null, "comment":'
-        ' null, "row_format": null, "fields_terminated_by": null, "lines_terminated_by": null,'
+        '"partitioned_by": [], "tablespace": null, "stored_as": "DIRECTORIES", '
+        '"row_format": null, "fields_terminated_by": null, "lines_terminated_by": null,'
         ' "map_keys_terminated_by": null, "collection_items_terminated_by": null, "external": false,'
-        ' "schema": null, "table_name": "list_bucket_single", '
+        ' "schema": null, "table_name": "list_bucket_single", "temp": false,'
         '"skewed_by": {"key": "key", "on": ["1", "5", "6"]}}]'
     )
     expected = json.loads(expected)
-    print(json.loads(parse_results))
-    print(expected)
     assert json.loads(parse_results) == expected
