+++ tests/test_simple_ddl_parser.py
@@ -818,7 +818,7 @@ def test_like_statement():
             "checks": [],
             "columns": [],
             "index": [],
-            "clone": {"schema": None, "table_name": "Old_Users"},
+            "like": {"schema": None, "table_name": "Old_Users"},
             "partitioned_by": [],
             "primary_key": [],
             "schema": None,
@@ -1902,7 +1902,7 @@ def test_quotes():
         `partition`   STRING,
     );
     """
-    parse_result = DDLParser(ddl).run(output_mode="hql")
+    parse_result = DDLParser(ddl).run()
     expected = [
         {
             "columns": [
@@ -1933,14 +1933,6 @@ def test_quotes():
             "index": [],
             "partitioned_by": [],
             "tablespace": None,
-            "stored_as": None,
-            "location": None,
-            "row_format": None,
-            "fields_terminated_by": None,
-            "lines_terminated_by": None,
-            "map_keys_terminated_by": None,
-            "collection_items_terminated_by": None,
-            "external": False,
             "schema": "`shema`",
             "table_name": "table",
             "if_not_exists": True,
@@ -1966,6 +1958,7 @@ def test_escaping_symbols_normal_str():
             {
                 "alter": {},
                 "checks": [],
+                "temp": False,
                 "collection_items_terminated_by": None,
                 "columns": [
                     {
@@ -2017,6 +2010,7 @@ def test_escaping_symbols_raw_string():
             {
                 "alter": {},
                 "checks": [],
+                "temp": False,
                 "collection_items_terminated_by": None,
                 "columns": [
                     {
@@ -2332,7 +2326,7 @@ def test_create_empty_table():
 
             CREATE TABLE "material_attachments"
     """
-    result = DDLParser(ddl).run(group_by_type=True, output_mode="hql")
+    result = DDLParser(ddl).run(group_by_type=True)
 
     expected = {
         "ddl_properties": [],
@@ -2343,19 +2337,11 @@ def test_create_empty_table():
             {
                 "alter": {},
                 "checks": [],
-                "collection_items_terminated_by": None,
                 "columns": [],
-                "external": False,
-                "fields_terminated_by": None,
                 "index": [],
-                "lines_terminated_by": None,
-                "location": None,
-                "map_keys_terminated_by": None,
                 "partitioned_by": [],
                 "primary_key": [],
-                "row_format": None,
                 "schema": None,
-                "stored_as": None,
                 "table_name": '"material_attachments"',
                 "tablespace": None,
             }
@@ -2372,7 +2358,7 @@ def test_table_name_reserved_word_after_dot():
 
     create table foo.index (col1 int);
         """
-    result = DDLParser(ddl).run(group_by_type=True, output_mode="hql")
+    result = DDLParser(ddl).run(group_by_type=True)
     expected = {
         "ddl_properties": [],
         "domains": [],
@@ -2382,7 +2368,6 @@ def test_table_name_reserved_word_after_dot():
             {
                 "alter": {},
                 "checks": [],
-                "collection_items_terminated_by": None,
                 "columns": [
                     {
                         "check": None,
@@ -2395,24 +2380,16 @@ def test_table_name_reserved_word_after_dot():
                         "unique": False,
                     }
                 ],
-                "external": False,
-                "fields_terminated_by": None,
                 "index": [],
-                "lines_terminated_by": None,
-                "location": None,
-                "map_keys_terminated_by": None,
                 "partitioned_by": [],
                 "primary_key": [],
-                "row_format": None,
                 "schema": None,
-                "stored_as": None,
                 "table_name": "index",
                 "tablespace": None,
             },
             {
                 "alter": {},
                 "checks": [],
-                "collection_items_terminated_by": None,
                 "columns": [
                     {
                         "check": None,
@@ -2425,24 +2402,16 @@ def test_table_name_reserved_word_after_dot():
                         "unique": False,
                     }
                 ],
-                "external": False,
-                "fields_terminated_by": None,
                 "index": [],
-                "lines_terminated_by": None,
-                "location": None,
-                "map_keys_terminated_by": None,
                 "partitioned_by": [],
                 "primary_key": [],
-                "row_format": None,
                 "schema": "foo",
-                "stored_as": None,
                 "table_name": "[index]",
                 "tablespace": None,
             },
             {
                 "alter": {},
                 "checks": [],
-                "collection_items_terminated_by": None,
                 "columns": [
                     {
                         "check": None,
@@ -2455,17 +2424,10 @@ def test_table_name_reserved_word_after_dot():
                         "unique": False,
                     }
                 ],
-                "external": False,
-                "fields_terminated_by": None,
                 "index": [],
-                "lines_terminated_by": None,
-                "location": None,
-                "map_keys_terminated_by": None,
                 "partitioned_by": [],
                 "primary_key": [],
-                "row_format": None,
                 "schema": "foo",
-                "stored_as": None,
                 "table_name": "index",
                 "tablespace": None,
             },
@@ -2679,7 +2641,6 @@ def test_increment_column():
                     "unique": False,
                 },
             ],
-            "constraints": {"checks": None, "references": None, "uniques": None},
             "index": [],
             "partitioned_by": [],
             "primary_key": ["user_id"],
@@ -2699,7 +2660,7 @@ def test_increment_column():
     );
     """
 
-    result = DDLParser(ddl).run(output_mode="mysql")
+    result = DDLParser(ddl).run()
 
     assert expected == result
 
@@ -2714,7 +2675,6 @@ def test_replace_with_id():
             {
                 "alter": {},
                 "checks": [],
-                "collection_items_terminated_by": None,
                 "columns": [
                     {
                         "check": None,
@@ -2727,26 +2687,17 @@ def test_replace_with_id():
                         "unique": False,
                     }
                 ],
-                "external": False,
-                "fields_terminated_by": None,
                 "index": [],
-                "lines_terminated_by": None,
-                "location": None,
-                "map_keys_terminated_by": None,
                 "partitioned_by": [],
                 "primary_key": [],
                 "replace": True,
-                "row_format": None,
                 "schema": None,
-                "stored_as": None,
                 "table_name": "someTable",
                 "tablespace": None,
-                "transient": True,
             },
             {
                 "alter": {},
                 "checks": [],
-                "collection_items_terminated_by": None,
                 "columns": [
                     {
                         "check": None,
@@ -2759,21 +2710,13 @@ def test_replace_with_id():
                         "unique": False,
                     }
                 ],
-                "external": False,
-                "fields_terminated_by": None,
                 "index": [],
-                "lines_terminated_by": None,
-                "location": None,
-                "map_keys_terminated_by": None,
                 "partitioned_by": [],
                 "primary_key": [],
                 "replace": True,
-                "row_format": None,
                 "schema": None,
-                "stored_as": None,
                 "table_name": "someTable",
                 "tablespace": None,
-                "temp": True,
             },
         ],
         "types": [],
@@ -2787,7 +2730,7 @@ def test_replace_with_id():
     );
     """
 
-    result = DDLParser(ddl).run(group_by_type=True, output_mode="hql")
+    result = DDLParser(ddl).run(group_by_type=True)
 
     assert expected == result
